Add --linelength command line option.
Adds a small but useful additional feature to hexedit: a mechanism to explicitly
specify the number of bytes to display per line. I find that a lot of binary
files that I look at had some sort of internal alignment that made them easier
to visualize if each line contains a specific number of bytes.