Ignore KEY_RESIZE in main loop.
The KEY_RESIZE key can be encountered while resizing the terminal
running hexedit.

As the resizing is handled in an own signal handler, and because
this key is not mapped, ignore it in key_to_function so no
error message is shown.

As a side note: The signal handler is not thread-safe, but that
is a subject for another patch.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>