Add a CI test for GitHub
It will run after each push to test an initial build, install and uninstall.
After this a distclean and a second build and install will run to test if
these actions execute twice. Any error in source code will break the build
and invalidate the CI test.

Currently, part of the test is disabled because the uninstall target doesn't
exist yet.